fix: change uint32 to float for price

+++ chips.hpp
@@ -16,7 +16,7 @@ class Chips : public Product
     Chips(const std::string & pn = "Null",
           const std::string & pc = "Null",
           uint32_t st = 0,
-          uint32_t pr = 0,
+          float pr = 0,
           float s = 0.0);
     ~Chips();
